Extract original implementation to bitmap module
Bitmap module contains original Roaring bitmap wrapper implementation.
To-be-created treemap module will contain bitmap which supports 64bit unsined
integers.
